Added Inventory in Admin
