--- nextjs_app/components/adminReports/inventory.jsx
@@ -1,20 +1,0 @@
-//this is related to Jojo's gagawin
-const inventory = () => {
-    const details= {
-        orderNo : 123456 ,
-        orderAmt: 1234,
-        paymentMtd: "GCASH"
-    }
-
-    return(
-        <div>
-            <div> 
-                  <h1 className="text-6xl"> Hello</h1>  
-            </div>
-        </div>
-
-        
-    )
-}
-
-export default inventory;
