+++ nextjs_app/components/reports/InventoryReport.jsx
@@ -1,81 +1,73 @@
-import ReportCard from "./ReportCard"
-
-
 const InventoryReport = ({month, year}) => {
-  //TODO: fetch data based on month and year props
-  const analytics = {
-    begBal: 200,
-    add: 100,
-    sales: 150,
-    others: 50,
-  }
 
   //TODO: fetch categories
-  const categories = [
+  const items = [
     {   
-        id: 1,
-        name: "Fruits",
-        qty: 100,
+        productImg:'/kamotePic.jpg',
+        name: "Kamote",
+        qty: "54 kg",
+        price: "120/kg",
     },
     {
-        id: 2,
-        name: "Vegetables",
-        qty: 200,
-    },
-    {
-        id: 3,
-        name: "Meat",
-        qty: 50,
+        productImg:'/chickenBreastPic.jpg',
+        name: "Chicken Breast",
+        qty: "62 kg",
+        price: "170/kg",
     }
-
   ]
   
     return (
-        <div className="bg-green-600 mb-6">
-            <div className="w-96 p-4 flex justify-center">
-                <table className="w-full">
+        <div>
+  
+            <div className="text-center font-serif color text-amber-400 m-4 text-4xl font-extrabold">
+                <h2> INVENTORY </h2>
+            
+              
+            </div>
+        
+     
+            <div className="overflow-x-auto text-black">
+                <table className="min-w-full">
+                    <thead>
+                        <tr className="text-center bg-amber-400 font-extrabold font-serif text-2xl">
+                        <th className="text-center">PRODUCT</th>
+                        <th className="text-center">PRODUCT NAME</th>
+                        <th className="text-center">QUANTITY</th>
+                        <th className="text-center">PRICE</th>
+                        <th className="text-center">OPERATION</th>
+                        </tr>
+                    </thead>
+                
                     <tbody>
-                        <tr>
-                            <td>Beginning Balance</td>
-                            <td>{analytics.begBal}</td>
-                        </tr>
 
-                        <tr>
-                            <td>Additions</td>
-                            <td>{analytics.add}</td>
-                        </tr>
-
-                        <tr>
-                            <td>Sales</td>
-                            <td>{analytics.sales}</td>
-                        </tr>
-
-                        <tr>
-                            <td>Others <hr/></td>
-                            <td>{analytics.others} <hr/></td>
-                        </tr>
-                        <tr>
-                            <td>Ending Balance</td>
-                            <td>{analytics.begBal + analytics.add - analytics.sales + analytics.others}</td>
-                        </tr>
-                    </tbody>
+            {items.map((item, index) => (
+              <tr key={index} className="py-4 text-center  items-center">
+                <td className="align-middle">
+                  {/* Items Display */}
+                  <img src={item.productImg} alt={item.name} width={100} height={100}  />
+                </td>
+                <td>{item.name}</td>
+                <td>{item.qty}</td>
+                <td>{item.price}</td>
+                <td className="items-center">
+                  
+                  <div className="space-x-2">
+                    <button className="px-2 py-1 bg-blue-500 text-white rounded">View</button>
+                    <button className="px-2 py-1 bg-red-500 text-white rounded">Delete</button>
+                  </div>
+                </td>
+              </tr>
+            ))}
+          </tbody>
+                    
+                        
                 </table>
             </div>
-            
-
-
-            <div>
-                <h2 className="font-extrabold text-white m-4 text-xl">Ending Inventory per Category</h2>
-                {categories.map((cat)=> (
-                    <ReportCard key={cat.id} title={cat.name} amount={cat.qty}/>
-                ))}
-            </div>
-
-            
-
-
-
+        
+                
+                
         </div>
+      
   )
 }
